Add MSP_RTC message
Returns the RTC from the FC to the sender. Note that MSP_SET_RTC
has been changed to 247 and MSP_RTC is now code 246. This follows
the convention of other MSP message pairs, where the getter has
a lower number than the setter.

Since there are still no MSP senders which send these messages,
there should be no problem with the code modification.