Revert bit order in bitarray
Unfortunately, bitarrays are used in several places where the data
is finally written to the wire, so reversing their order would
mean breaking compatibility with software expecting those bits in
the reverse order.

This has the consequence to not allow us to use clz directly in
bitArrayFindFirstSet, requiring an intermediate rbit instruction
which adds around 1ns time penalty to max7456DrawScreenPartial().

Fixes #2154