Test for GPS enabled rather than GPS present to draw GPS indicators
Drawing it only when GPS is present would cause the GPS related
information to disappear from the screen if the GPS goes bad
during a flight. This would erase the coordinates from the screen
and might difficult a successful recovery of and aircraft. It can
also be confusing while setting up a new model.