Update navigation_multicopter.c
Issue #1793 Activating AltHold While Inverted:
If inverted when activating ALT-Hold, reverse motors then at the top of the flip change motors to the normal direction and reset the hold position to the current position. This is a big improvement (see video link under issue #1793) and I have made some more progress since the video was made.  More work such as first level-out, punch-out, then flip would be beneficial.  Event just leveling the pitch before the flip would smooth out the transition.  After this is perfected, it would be great to add other exciting features such as hard-deck to automatically activate this feature when a certain altitude is reached making it nearly impossible to crash.  This would be a revolutionary improvement to our hobby that would re-establish 3D Quads/Helicopter-modes that would attract and bring back many people to this exciting hobby.