Update mixer.c
Issue #10879 3D Not Symmetrical Issued:
As a work-around, I changed the 3D Neutral variable to use the Configurator-[3D Neutral] value to adjust the center stick transition point.

I also changed the [Deadband Low] and [Deadband High] defaults as the existing values are too close to each other and do not work well.  These values keep the motors spinning faster in the Deadband zone, which make the quad much smoother and prevents the severe jitters when the throttle is reapplied. 

For 3D the configuration setting [hover throttle] needs to be around 1650 as the default 1500 is zero throttle in 3D mode.