Use characters with trailing/leading half-dot rather than trailing dot
Having the whole dot in one character made it look bad in some
instances, since the character with the dot had to be moved several
pixels to the left. Having half a dot on each characters allows us to
keep the numbers at the same location inside the character area making
it look more balanced.