Run the settings generator only when the build changes
A build is considered changed if either the TARGET, CFLAGS or
the #define'd constants by including target.h and platform.h change

The Makefile now runs a new tool in src/utils/build_stamp.rb which
updates the file in obj/build.stamp iff the build has changed. This
allows us to make the generated files to depend on this stamp file
rather than rebuilding them all the time. Automatic dependencies
generated by the compiler take care of recompiling just the files
which depend on the generated ones.