Use typed functions/enums rather than macros for setting flags
This allows the compiler to check switches handling those values
and make sure all cases are handled. Would have prevented the bug
fixed in 966e97c.

Note that the generated assembly code is exactly the same, we're
just giving the compiler more information in order to be able to
detect more errors for us.