Add OSD alarms for distance and negative altitude
Distance alarm makes the OSD blink when distance to home goes
above a given threshold. Default is 1000m and it always works
on meters to avoid performing unnecessary arithmetic operations
for displaying the distance indicator. Unit conversion is done
in the configurator instead. Zero disables the alarm.

Negative altitude alarm blinks when negative altitude is below
the given threshold. Note that the alarm is a positive number,
but will trigger with -ALT > NEG_ALT_ALARM. Default value is 5
(which means -5 meters). Again, conversion is pushed to the
configurator.

Fixes #2242 and #2199