Initial watchdog support
For now, it just restarts the FC if it gets stuck for more
than 500ms.
The watchdog is only called from the system task in scheduler.c
right now. Eventually we might need to call it from more places.
Also, flashing a new firmware leaves the processor stuck and requires
a power cycle to make it work again. We might to adjust how we
restart into DFU mode with a WD present in order to avoid the
power cycle requirement.