Replace two battery capacity functions with just one
Remove calculateBatteryPercentage() and
calculateBatteryCapacityRemainingPercentage(), replacing them with
calculateBatteryPercentageRemaining() which uses the battery
capacity when it's available and current sensor is enabled, but
falls back to estimation based on voltage otherwise.

This means that we no longer have two battery bars in the dashboard,
so the second one has been removed.