Update documentation for building in Mac OS X
As all new binary and source packages for GNU ARM Embedded Toolchain will not be released on Launchpad henceforth, a change is required for this documentation to be up to date. The compiler version is also too old (was 2015 release) and need to be updated. The example featuring 'NAZE' is also updated to 'SPRACINGF3'. The expired information in this documentation caused confusion in https://github.com/iNavFlight/inav/issues/2543