Highly experimental support for using GPS heading on MR
Uses the same logic as FW GPS heading, but applies a correction to
the heading reported by the GPS by using the tilt angle of the MR
head.

Please, be very careful testing this feature for the moment.

This commit introduces the new setting use_gps_heading, with
the following values:

- AUTO (default): Enabled just for FW
- ON: Enabled for all crafts
- OFF: Disabled for all crafts