Don't show the heading in the OSD if we don't know it
While no heading is know, the graph indicator shows an unlabeled
graph, while the numerical indicator will show "---º".

This is useful both for crafts with MAG (it possible to see if
MAG went bad after arming) and for crafts using GPS based heading
(it lets the pilot know when the heading has been initialized).