Move CMS entries and menus from RAM to FLASH
Use a global array in cms.c to keep track of the dirty elements
that need redrawing. This lets us avoid modifying the entry flags
to keep track of it, so they can now be completely read only. This
in turn lets us also move all the menus to the text section.

This commit saves 6112 bytes of RAM in F3, so we have some room
to play again without resorting to disabling features.