Send GPS related sensors via S.Port even without a fix
Check for feature(FEATURE_GPS) rather than for a GPS fix. This lets
users perform the sensor discovery step without having a GPS fix.

To still allow users to notice if the GPS has stopped working, we
check that either the craft has never been armed or we do have a GPS
fix. This way the sensor is available on startup, but if it stops
working mid-flight the user will get a "sensor lost" alarm.