Fix maps with reference heading != 0º
The reference heading rotation was reversed, but it didn't affect
maps oriented north or when the takeoff direction was close to 0º.

As a result, radar mode had an incorrect extra rotation applied to
it to compensate for this incorrect rotation in the map heading.
Since that's now fixed, the extra rotation is now removed from
osdDrawRadar().