Override the OSD layout for 10 seconds when configuring an item
When configuring the OSD via MSP (e.g. from the configurator), the
layout that was just altered is shown in the OSD for 10 seconds,
regardless of the selected OSD layout via modes.

This allows configuring the layouts when looking at them in the
goggles while the radio is powered off.