Make the OSD home arrow closer to the actual home direction
We were using the first image (up) for [0deg, 22deg] which made
some directions like [350deg, 360deg) feel pretty unnatural because
the craft was almost pointing home but the arrow pointed a bit to
the left. Maximum error between actual direction and the arrow was
22deg.

With this change, the up image is used for [349deg, 11deg], 2nd
image covers [12deg, 33deg], etc... which makes the error between
the actual direction and the arrow smaller. This also cuts the
maximum error between the arrow and the actual direction to ~12deg.