Teach settings how to handle strings, make craft name a setting
Add a new type for settings to be able to handle strings. This makes
it easier to add more settings of this type and removes the need to
handle them with dedicated CLI commands.

Replace the `name` CLI command with a setting called `name`.

Saves 8 bytes of FLASH on OMNIBUS