Teach the map drawing code to reduce scale
When looking for another scale when drawing would overwrite a
non-blank character, reduce the scale when we're close to the center
and increase it when we're far from it. Otherwise there are some
pathological cases when an indicator next to the map center would
cause the map to not draw anything.