Make the home arrow show only when we know all needed data
Home arrow will only be displayed if we have a GPS fix, we have
a home direction recorded and we have a proper heading. Otherwise
the indicator will blink and it will show a fixed arrow point up
while unarmed (so users can see the arrow when configuring the OSD)
and a '-' character when armed (so users can see if some sensor
has failed during a flight).