Fix detection of NTSC signal after 1.5s
Due to a logic error during refactor, signal was always being stored
internally as PAL when more than 1.5s had elapsed since boot. The
MAX7456 kept reporting that the signal was actually NTSC, causing the
reinitialization sequence to trigger and the OSD to blink once every
second, but since the data cached in RAM was always stored as PAL
this continued forever.

Fixes #4183