Convert servos to use [-1, 1] internally
Servos subsytem now works internally with floats using [-1, 1] as
its domain. Values are converted to PWM when calling pwmWrite().
Some operations have also been refactored to reduce the number
of instructions and branches executed during the main PID loop,
reducing the time needed for servos on an airplane with flaps
from 15us to 13us on F4.

+242 flash, +128 ram