Revert "[PG] Avoid keeping 2 copies of each PG in RAM"
Unfortunately, making this work properly requires a lot of changes
to the targetConfiguration() system, but still leave us with all
the shortcomings of the PGs.

Working on a new and better way to store the settings seems like
a more productive use of time.

On the bright side, this uncovered a couple of bugs in the
`diff` logic that I will fix on the next commit.

Fixes #5458