[SYSTEM] Avoid performing a 64bit division in micros()
Since the sub-ms step fits in an uint32_t, we can do the calculation
using just 32 bits, then perform a 64 bit add to the (ms * 1000)
value. This is way faster because cortex-m[4,7] do have 32 bit
integer division in HW, but not for 64 bit.

This reduces the number of cycles used by micros() from ~177
to ~57 (there's a small 3-4 cycle variance due to pipelining).