[MAKE] Avoid adding our custom CFLAGS twice
We don't clear CFLAGS becase we want to allow them to be passed
via command line. However, when building a target we add our
custom ones once in the first parsing of the Makefile and then
again a second time when we invoke make recursively for each target.
This breaks compilation with -Werror.

To workaround this, save the initial CFLAGS in a variable, then
pass them to the recursive invocation of make.