[OSD] Allow moving sidebars left and right
Introduce a osd_sidebar_horizontal_offset parameter that uses
grid slots or pixels as units, depending on the OSD type. int8_t
is enough to move the sidebars to the corner of the screen even
with pixel OSD. Positive values move the bars closer to the screen
edges, while negative values bring them closer to the center.