[RX] Make halfDuplex (exposed as serialrx_halfduplex) a tristate
This allows us to set the default value to AUTO, so each protocol
can select its default "duplexity". Then the explicit ON and OFF
values can be used to override the default value for the protocol.

Regenerate Settings.md to update the serialrx_halfduplex
documentation. This also regenerates some unrelated settings that
were out of date.

Update Rx.md accordingly.