Returns GCS waypoint in the case of waypoint 255
For MSP query in the case of waypoint 255, in this commit the ground control station waypoint is returned, which is consistent with its waypoint set behavior. This is also the desired behavior to affirm users that the correct waypoint is loaded. If GCS assist mode is not enabled, the current position is returned to preserve compatibility.