Updated UK units to be correct for the UK
The units used as "UK" were incorrect. We use imperial for distance and speed. To correct this, I have added a new units set called Metric + MPH.

Metric + MPH is the old UK units set. Anyone who used the old UK setting would use Metric + MPH going forward. As in, if you had UK before, did the update, you would now be on Metric + MPH. The transition is seamless to the pilot. This is due to adding it before UK in the osd_unit_e enum.

UK is effectively the new units set. It uses imperial for everything, except the temperature which is in degrees C.