Updated after initial tests
I performed some initial testing of the GA units today. The distances in NM are correct, as is the speed in Knots. The knots symbol needs to be changed though. The numeric vario was not working, but this should now be correct. I will test again soon.

However, during testing I noticed that the conversion to feet was incorrect. At 200 metres, it was out by 3 feet. So this is not an insignificant error. This should now be fixed.