Fixing the missing commands to ESC
After using the turtle mode, depending on the configuration of the hardware, not all commands are apparently completely transferred to the ESCs. This means that some motors still have the wrong direction of rotation, even though the turtle mode has been deactivated. This queue addresses these issues.
The mentioned error has thus been eliminated
and i couldn't see any disadvantages, but i'm not a developer and grateful for any other better solution.