Remove unnecessary parameters from profiles
`fw_loiter_direction` was in control profiles. There's no reason why this would want to be changed with different control parameters. It's purely a pilot preference. This has been moved to `navConfig.fw`

'fw_min_throttle_down_pitch` was in the battery profiles. This command is really something that is set based on the airframe and it's glide ratio. Battery type or performance does not effect this. Moved to `navConfig.fw`