Increase nav_wp_safe_distance
This PR increases the maximum safe distance for the first waypoint to 1500m. A lot of fixed wing LR pilots are frustrated with the 650m limit. This corrects that while not increasing the memory used by the variable. I didn't see any reason for this to be in cm. It's not divided anywhere. So metres is much more fitting.