Change number of digits in battery voltage display.
In BF compatible mode the . takes up space of data that usually gets
displayed.

It i possible to work around it for the main battery voltage by changing
from 1 decimal digits to 2, but there will always only be space for N -
1  digits.

This bumps the size of the field by 1 in BF_COMPATIBLE mode for main
battery voltage and cell voltage fields.