Additional updates to support HDZERO.
Updated the function which decides whether to support single page stats, osdVideoSystemIsSinglePageStatsCompatible(), to be based on the enumerated value for the selected video system.

Combined MAX SPEED and AVG SPEED on to one line.

Combined MIN RSSI % and MIN RSSI DBM on to one line.

So we're down to 16 lines, 17 including the blank line at the top.