Fix LED Indicator mode for fixed wing
Currently, the indicator mode is only really works for quads. It doesn't make any sense for fixed wing. This PR changes that. Roll movements to the left triggers indicators on the left half of the grid. Roll movements to the right triggers indicators on the right half of the grid.

This change may also be suitable for rovers too. If so, it's an easy modification. So feedback from rover people would be great.