Fix SITL compilation on macosx (#9012)
Clean up cmake files for SITL and fix the compilation issues caused by tooling differences between gcc and clang (used in MacOSX).

SITL on MacOSX still has lots of compile time warnings that need to be addressed and some known issues that prevent it from working with simulators, but those will be addressed in 7.0.

At this point, I don't think MacOSX SITL is ready to be added to the configurator. That should be targeted for 7.0.