Enable 10Hz update rate for M9+ gps devices
Update rate of M9 devices was limited to 5Hz to enable tracking of more
than 16 satellites.

The number of tracked satellites is limited by the update rate in M8
devices as well, but we allow those to run in 10Hz mode.

This will change the behavior of M9 and newer devices to match M5-8,
enabling 10Hz when using UBLOX7 and limiting it to 5Hz when using
regular UBLOX.