Only show rearm message when not saving
If the save is happening, writing to EEPROM locks up the FC. This change will only show the rearm message if the FC is not currently saving.

If rearm is available, the `** SETTINGS SAVED**` message will not appear, to maximise time for rearm.

If rearm is not available. The save messages display as normal.