Change SmartPort sensor IDs for GPS and Modes
The current implementation of INAV SmartPort telemetry uses temperature sensor IDs for Modes and GNSS data. This causes issues on transmitter firmware that more strictly adhere to the telemetry ID's intended use. The Tmp sensors can't be converted to a different type, so they can't be read correctly.

This PR changes the IDs to some that are not associated with specific functions or types. There is also an option to use the legacy IDs to allow for a deprecation period. This legacy option should be removed in INAV 10.0.