Add LYIF722D TARGETS
The LYIF722D is a flight controller that uses the STM32F722RGT6 MCU and has dual gyroscopes. this is an unofficial target .The input voltage ranges from 7V to 40V, which is equivalent to 2-8S. It has BECs of 5V 3A and 12V 3A, and it comes with an OSD, a barometer, and 1-8 DSHOT motor outputs.I will provide pictures, definitions and other content related to the corresponding products within the topics discussed below.