Update osd.c - RSSI Display Code
The current way the RSSI is displayed is such that the RSSI takes up two characters. Whilst this is OK for most OSD systems, it causes alignement issues with regards to the values when used with LQ, which hovers around 100 most of the time. 

I think that the MAX symbol should not be used, but instead the value should be shown as 100 when it is 100 same as the LQ. This way both the LQ and the RSSI, which are most frequently used together will both be showing with the same total number of characters, i.e. 4 in total, 1 for the signal bars icon and 3 for the value. This way users can correclty align the two signal icons one under the other and the values can lign up to the right. 

The only thing that would need to be done would be to adjust the lenght in the configurator as well.