DJI Strikes again
The O4 firmware currently shipping has a regression, when compared to
DJI O3.

They are not accepting INAV's reply to STATUS/STATUS_EX commands.

This patch adds the option to re-use the workaround needed for the
Original DJI FPV System, where INAV would reply with a custom message to
those messages.

The main difference is related to the number of flight modes in INAV vs
Betaflight. INAV has a lot more modes and is likely sending a reply that
DJI thinks is too big and marked as invalid.

While the arming bit is supposed to be the same accross BF and INAV,
they are also not using the correct workflow of asking the firmware for
the list of modes supported (and corresponding bits) to check what is
active or not.

While it is not critical to check the modes, as it is very unlikely the
ARM bit will move/change, it is certainly incorrect to ignore the reply
based on the number of modes, if you haven't checked how many modes
there should be.