Add scaling to pan servo offset
This PR adds scaling to the pan servo offset. Historically it based the degree range on a typical 1000-2000 microseconds pulse width. However, many servos need to extend that pulse width to actually get the desired output from the servo. The `osdGetPanServoOffset` function will now automatically scale the degrees to the min/max range the servo is set to use.

Changes:
- Add scaling to `osdGetPanServoOffset`
- Make the docs easier to read
- Renamed the parameter, as centi-degrees was completely the wrong term. Now deca-degrees

Requires Configurator update