Update Geofence `AUTO` mode to `GEO`
When the Geofence automated flight is active, it currently says `AUTO` in the OSD. It would be nicer if it were a bit more specific as to what is happening. So now will say `GEO`. This is also reflected in the CRSF telemetry and will be reflected in S.Port telemetry too,