Update references to that should be control_profile
This is to clean up some old references that should have been `control_profile`. But, have remained as different options.

Within the firmware, there are `controlProfile` (previously `controlRateProfile`) and `PIDProfile`. These are separate entities internal to the firmware. But, to the user, they only see `control_profile` which contains all the settings included within the firmware's separate `controlProfile` and `PIDProfile`.

This PR aims to take confusion away from the user by just referencing `control_profile`. This was most noticeable in the automated `control_profile` switching available in the `battery_profile` and the `mixer_profile`. Neither of which reference the UI visible `control_profile`. But mentioned `controlrate_profile` and `PID Profile` respectively. Neither of which are separately defined in the UI or CLI.

In future, it is worth remembering that the user sees all `PIDProfile` and `control_profile` settings within the `control_profile` in CLI or as a selectable **Control Profile** in Configurator.